Conflicts:
	ambry-coordinator/src/test/java/com.github.ambry.coordinator/CoordinatorTest.java
	ambry-coordinator/src/test/java/com.github.ambry.coordinator/OperationPolicyTest.java

+++ ambry-coordinator/src/test/java/com.github.ambry.coordinator/CoordinatorTest.java
@@ -889,24 +889,29 @@ public class CoordinatorTest {
       throws JSONException, InterruptedException, StoreException, IOException, CoordinatorException {
 
     for (ServerErrorCode deleteErrorCode : deleteErrorMappings.keySet()) {
-      MockConnectionPool.mockCluster = new MockCluster(clusterMap);
-      induceDeleteFailure(TOTAL_HOST_COUNT, deleteErrorCode);
-      AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
-      for (int i = 0; i < 20; ++i) {
-        BlobProperties putBlobProperties =
-            new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
-        ByteBuffer putUserMetadata = getByteBuffer(10, false);
-        ByteBuffer putContent = getByteBuffer(100, true);
-        String blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
-        getBlob(ac, blobId, putBlobProperties, putUserMetadata, putContent);
-        try {
-          deleteBlobAndGet(ac, blobId);
-          Assert.fail("Deletion should have failed for " + deleteErrorCode);
-        } catch (CoordinatorException e) {
-          Assert.assertEquals(e.getErrorCode(), deleteErrorMappings.get(deleteErrorCode));
+      if (!deleteErrorCode.equals(ServerErrorCode.Disk_Unavailable) &&
+          !deleteErrorCode.equals(ServerErrorCode.Partition_ReadOnly) &&
+          !deleteErrorCode.equals(ServerErrorCode.IO_Error)) {
+        // Ignoring three ServerErrorCodes as PUTs are expected to fail in such cases
+        MockConnectionPool.mockCluster = new MockCluster(clusterMap);
+        induceDeleteFailure(TOTAL_HOST_COUNT, deleteErrorCode);
+        AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
+        for (int i = 0; i < 20; ++i) {
+          BlobProperties putBlobProperties =
+              new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
+          ByteBuffer putUserMetadata = getByteBuffer(10, false);
+          ByteBuffer putContent = getByteBuffer(100, true);
+          String blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
+          getBlob(ac, blobId, putBlobProperties, putUserMetadata, putContent);
+          try {
+            deleteBlobAndGet(ac, blobId);
+            Assert.fail("Deletion should have failed for " + deleteErrorCode);
+          } catch (CoordinatorException e) {
+            Assert.assertEquals(e.getErrorCode(), deleteErrorMappings.get(deleteErrorCode));
+          }
         }
+        ac.close();
       }
-      ac.close();
     }
   }
 
@@ -939,23 +944,29 @@ public class CoordinatorTest {
       throws JSONException, InterruptedException, StoreException, IOException, CoordinatorException {
 
     for (ServerErrorCode putErrorCode : putErrorMappings.keySet()) {
-      MockConnectionPool.mockCluster = new MockCluster(clusterMap);
-      inducePutFailure(TOTAL_HOST_COUNT, putErrorCode);
-      AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
-      for (int i = 0; i < 20; ++i) {
-        BlobProperties putBlobProperties =
-            new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
-        ByteBuffer putUserMetadata = getByteBuffer(10, false);
-        ByteBuffer putContent = getByteBuffer(100, true);
-        String blobId = null;
-        try {
-          blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
-          Assert.fail("Put should have failed for " + putErrorCode);
-        } catch (CoordinatorException e) {
-          Assert.assertEquals(e.getErrorCode(), putErrorMappings.get(putErrorCode));
+      System.out.println("puError code " + putErrorCode);
+      if (!putErrorCode.equals(ServerErrorCode.Disk_Unavailable) &&
+          !putErrorCode.equals(ServerErrorCode.Partition_ReadOnly) &&
+          !putErrorCode.equals(ServerErrorCode.IO_Error)) {
+        // Ignoring three ServerErrorCodes as PUTs are expected to fail in such cases
+        MockConnectionPool.mockCluster = new MockCluster(clusterMap);
+        inducePutFailure(TOTAL_HOST_COUNT, putErrorCode);
+        AmbryCoordinator ac = new AmbryCoordinator(getVProps(), clusterMap);
+        for (int i = 0; i < 20; ++i) {
+          BlobProperties putBlobProperties =
+              new BlobProperties(100, "serviceId", "memberId", "contentType", false, Utils.Infinite_Time);
+          ByteBuffer putUserMetadata = getByteBuffer(10, false);
+          ByteBuffer putContent = getByteBuffer(100, true);
+          String blobId = null;
+          try {
+            blobId = putBlob(ac, putBlobProperties, putUserMetadata, putContent);
+            Assert.fail("Put should have failed for " + putErrorCode);
+          } catch (CoordinatorException e) {
+            Assert.assertEquals(e.getErrorCode(), putErrorMappings.get(putErrorCode));
+          }
         }
+        ac.close();
       }
-      ac.close();
     }
   }
